Change app name Update Junit tests Added guihandles: TodoCountHandle
