+++ src/main/java/seedu/address/commons/core/Config.java
@@ -11,7 +11,7 @@ public class Config {
     public static final String DEFAULT_CONFIG_FILE = "config.json";
 
     // Config values customizable through config file
-    private String appTitle = "Address App";
+    private String appTitle = "AcquaiNote";
     private Level logLevel = Level.INFO;
     private String userPrefsFilePath = "preferences.json";
