+++ src/test/java/seedu/address/commons/core/ConfigTest.java
@@ -14,7 +14,7 @@ public class ConfigTest {
 
     @Test
     public void toString_defaultObject_stringReturned() {
-        String defaultConfigAsString = "App title : Address App\n"
+        String defaultConfigAsString = "App title : AcquaiNote\n"
                 + "Current log level : INFO\n"
                 + "Preference file Location : preferences.json";
