+++ src/test/java/seedu/address/commons/util/AppUtilTest.java
@@ -15,7 +15,7 @@ public class AppUtilTest {
 
     @Test
     public void getImage_exitingImage() {
-        assertNotNull(AppUtil.getImage("/images/address_book_32.png"));
+        assertNotNull(AppUtil.getImage("/images/AcquaiNote.png"));
     }
